Move wmi fetch to util function
We can't use the diag service for this since it runs before app launch
